report transaction today

+++ database/migrations/2025_04_17_025008_create_produks_table.php
@@ -18,9 +18,9 @@ return new class extends Migration {
             $table->foreignId('id_kategori')->constrained('kategori')->onDelete('cascade')->onUpdate('cascade');
             $table->integer('stok_masukSementara')->default(0);
             $table->integer('stok_akhirSementara')->default(0);
-            $table->decimal('hpp', 10, 2)->nullable();
-            $table->decimal('harga', 10, 2);
-            $table->decimal('diskon', 10, 2)->nullable();
+            $table->integer('hpp')->nullable();
+            $table->integer('harga');
+            $table->integer('diskon')->nullable();
             $table->timestamps();
         });
     }
